# voxd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "voxd",
    "description": "\nThe local API behind VoxStudio. Everything the app does, your scripts can do too.\n\n**Authentication:** send `Authorization: Bearer <token>` on every request except\n`GET /v1/status`. Audio URLs used directly in `<audio>` tags may pass `?token=<token>` instead.\n\n**Errors** always look like `{\"error\": \"<code>\", \"message\": \"<human text>\"}`.\n\n**Long work** runs as a job: queue it, then follow `GET /v1/jobs/{id}/events` (server-sent events)\nor the app-wide `WS /v1/events` stream.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/speech": {
      "post": {
        "tags": [
          "Speech"
        ],
        "summary": "Generate speech",
        "description": "Render short text (up to 5,000 characters) and wait for the result. For longer text use `POST /v1/jobs/speech`.",
        "operationId": "speech_v1_speech_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakeOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/transcriptions": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Transcribe an audio or video file",
        "description": "Starts a transcription job. On success its `result.transcript_id` points to the transcript.",
        "operationId": "create_transcription_v1_transcriptions_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_transcription_v1_transcriptions_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transcription"
        ],
        "summary": "List transcripts",
        "operationId": "list_transcriptions_v1_transcriptions_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscriptSummaryOut"
                  },
                  "title": "Response List Transcriptions V1 Transcriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audio/speech": {
      "post": {
        "tags": [
          "OpenAI compatible"
        ],
        "summary": "Create speech (OpenAI compatible)",
        "description": "Same request as OpenAI's `POST /v1/audio/speech`; returns audio bytes. The result is also saved as a take in History.",
        "operationId": "create_speech_v1_audio_speech_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The audio file",
            "content": {
              "audio/mpeg": {},
              "audio/ogg": {},
              "audio/aac": {},
              "audio/flac": {},
              "audio/wav": {},
              "audio/L16": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio/transcriptions": {
      "post": {
        "tags": [
          "OpenAI compatible"
        ],
        "summary": "Create transcription (OpenAI compatible)",
        "description": "Same request as OpenAI's `POST /v1/audio/transcriptions`. The transcript is also saved in Transcribe.",
        "operationId": "create_transcription_v1_audio_transcriptions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_transcription_v1_audio_transcriptions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_transcription_v1_audio_transcriptions_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Audio or video: flac, mp3, mp4, mpeg, mpga, m4a, ogg, wav, webm and more"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "`whisper-1`, `gpt-4o-transcribe` etc. use your preferred Whisper model; or a voxd Whisper model id",
            "default": "whisper-1"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "ISO-639-1 code; omit to detect"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "description": "Accepted for compatibility; ignored"
          },
          "response_format": {
            "type": "string",
            "enum": [
              "json",
              "text",
              "srt",
              "verbose_json",
              "vtt"
            ],
            "title": "Response Format",
            "default": "json"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "description": "Accepted for compatibility; ignored",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_transcription_v1_audio_transcriptions_post"
      },
      "Body_create_transcription_v1_transcriptions_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Any common audio or video format (wav, mp3, m4a, mp4, mov, webm\u2026), up to 2 GB"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "ISO code such as `en`; omit to detect automatically"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Whisper model id; defaults to your preference or the most accurate installed"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_transcription_v1_transcriptions_post"
      },
...
      "ErrorOut": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Job type, e.g. `speech`"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "queued | running | succeeded | failed | cancelled"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "description": "0.0 \u2013 1.0"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable current step"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Set when status is `succeeded`; shape depends on `kind`"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Set when status is `failed`"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "JobOut"
      },
...
      "SpeechIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "examples": [
              "Hello from VoxStudio."
            ]
          },
          "voice": {
            "type": "string",
            "title": "Voice",
            "description": "Voice id from GET /v1/voices",
            "examples": [
              "Samantha"
            ]
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine",
            "description": "Engine id; defaults to the first available TTS engine"
          },
          "speed": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.5,
            "title": "Speed",
            "description": "Playback rate multiplier",
            "default": 1.0
          },
          "emotion": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion",
            "description": "Emotional intensity, 0 (flat) \u2013 1 (dramatic); 0.5 is natural. Used by engines with the `emotion` capability."
          }
        },
        "type": "object",
        "required": [
          "text",
          "voice"
        ],
        "title": "SpeechIn"
      },
...
      "SpeechRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "`tts-1`, `tts-1-hd` or `gpt-4o-mini-tts` pick the best installed engine; or name one: `kokoro`, `chatterbox`, `system`",
            "default": "tts-1"
          },
          "input": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Input",
            "description": "Text to speak. Long text is rendered in chunks."
          },
          "voice": {
            "type": "string",
            "title": "Voice",
            "description": "An OpenAI voice name (alloy, ash, ballad, coral, echo, fable, nova, onyx, sage, shimmer, verse) or any voxd voice id",
            "default": "alloy"
          },
          "response_format": {
            "type": "string",
            "enum": [
              "mp3",
              "opus",
              "aac",
              "flac",
              "wav",
              "pcm"
            ],
            "title": "Response Format",
            "default": "mp3"
          },
          "speed": {
            "type": "number",
            "maximum": 4.0,
            "minimum": 0.25,
            "title": "Speed",
            "description": "Clamped to 0.5\u20132.0",
            "default": 1.0
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions",
            "description": "Accepted for compatibility; ignored"
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "SpeechRequest"
      },
...
      "TakeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "voice": {
            "type": "string",
            "title": "Voice"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "starred": {
            "type": "boolean",
            "title": "Starred"
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "description": "Unix timestamp (seconds)"
          },
          "audio_url": {
            "type": "string",
            "title": "Audio Url",
            "description": "Relative URL of the WAV file"
          }
        },
        "type": "object",
        "required": [
          "id",
          "engine",
          "voice",
          "text",
          "duration_s",
          "starred",
          "created_at",
          "audio_url"
        ],
        "title": "TakeOut"
      },
...
      "TranscriptSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "`file` or `live`"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "preview": {
            "type": "string",
            "title": "Preview",
            "description": "The first ~160 characters"
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "language",
          "duration_s",
          "model",
          "preview",
          "has_audio",
          "created_at"
        ],
        "title": "TranscriptSummaryOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}